refactor: add backlink to application components exports
