+++ pages/persoonsgegevens/succes.tsx
@@ -17,12 +17,12 @@ import {
   PageFooter,
   PageHeader,
   Paragraph,
+  ReservationCard,
   SkipLink,
   Surface,
 } from "../../src/components";
 import { PageFooterTemplate } from "../../src/components/huwelijksplanner/PageFooterTemplate";
 import { PageHeaderTemplate } from "../../src/components/huwelijksplanner/PageHeaderTemplate";
-import { ReservationCard } from "../../src/components/huwelijksplanner/ReservationCard";
 import { exampleState } from "../../src/data/huwelijksplanner-state";
 
 export const getServerSideProps = async ({ locale }: { locale: string }) => ({
