+++ pages/voorgenomen-huwelijk/index.tsx
@@ -6,6 +6,7 @@ import { useTranslation } from "next-i18next";
 import { serverSideTranslations } from "next-i18next/serverSideTranslations";
 import {
   Aside,
+  BackLink,
   ButtonGroup,
   ButtonLink,
   Document,
@@ -18,14 +19,13 @@ import {
   PageFooter,
   PageHeader,
   Paragraph,
+  ReservationCard,
   SkipLink,
   Surface,
   Link as UtrechtLink,
 } from "../../src/components";
-import { BackLink } from "../../src/components/BackLink";
 import { PageFooterTemplate } from "../../src/components/huwelijksplanner/PageFooterTemplate";
 import { PageHeaderTemplate } from "../../src/components/huwelijksplanner/PageHeaderTemplate";
-import { ReservationCard } from "../../src/components/huwelijksplanner/ReservationCard";
 import { exampleState } from "../../src/data/huwelijksplanner-state";
 
 export const getServerSideProps = async ({ locale }: { locale: string }) => ({
