+++ pages/voorgenomen-huwelijk/partner/uitnodigen.tsx
@@ -15,11 +15,9 @@ import { useRouter } from "next/router";
 import { useTranslation } from "next-i18next";
 import { serverSideTranslations } from "next-i18next/serverSideTranslations";
 import { FormEvent } from "react";
-import { ButtonGroup, ButtonLink, PageContentMain, PartnerInvitation } from "../../../src/components";
-import { BackLink } from "../../../src/components/BackLink";
+import { BackLink, PageContentMain, PartnerInvitation, ReservationCard } from "../../../src/components";
 import { PageFooterTemplate } from "../../../src/components/huwelijksplanner/PageFooterTemplate";
 import { PageHeaderTemplate } from "../../../src/components/huwelijksplanner/PageHeaderTemplate";
-import { ReservationCard } from "../../../src/components/huwelijksplanner/ReservationCard";
 import { exampleState } from "../../../src/data/huwelijksplanner-state";
 
 export const getServerSideProps = async ({ locale }: { locale: string }) => ({
