+++ src/components/index.ts
@@ -11,6 +11,7 @@ export * from '@gemeente-denhaag/process-steps';
 // Components for this project:
 export { Address } from './Address';
 export { Aside } from './Aside';
+export { BackLink } from './BackLink';
 export { ButtonGroup } from './ButtonGroup';
 export { Checkbox2 } from './Checkbox2';
 export { DataNoTranslate } from './DataNoTranslate';
